# util/mon,colexec/colexecdisk: reduce allocations