--- a/pkg/sql/colexec/colexecdisk/disk_spiller.go
+++ b/pkg/sql/colexec/colexecdisk/disk_spiller.go
@@ -90,10 +90,9 @@ func NewOneInputDiskSpiller(
 		diskBackedReuseMode:     diskBackedReuseMode,
 	}
 	op.diskSpillerBase = diskSpillerBase{
-		inputs:     []colexecop.Operator{input},
-		inMemoryOp: inMemoryOp,
-		// TODO(mgartner): Do not convert the name to a string.
-		inMemoryMemMonitorNames: []string{inMemoryMemMonitorName.String()},
+		inputs:                  []colexecop.Operator{input},
+		inMemoryOp:              inMemoryOp,
+		inMemoryMemMonitorNames: []mon.Name{inMemoryMemMonitorName},
 		diskBackedOpConstructor: op.constructDiskBackedOp,
 		spillingCallbackFn:      spillingCallbackFn,
 	}
@@ -177,15 +176,10 @@ func NewTwoInputDiskSpiller(
 	op := &twoInputDiskSpiller{
 		diskBackedOpConstructor: diskBackedOpConstructor,
 	}
-	names := make([]string, len(inMemoryMemMonitorNames))
-	for i := range names {
-		// TODO(mgartner): Do not convert the names to strings.
-		names[i] = inMemoryMemMonitorNames[i].String()
-	}
 	op.diskSpillerBase = diskSpillerBase{
 		inputs:                  []colexecop.Operator{inputOne, inputTwo},
 		inMemoryOp:              inMemoryOp,
-		inMemoryMemMonitorNames: names,
+		inMemoryMemMonitorNames: inMemoryMemMonitorNames,
 		diskBackedOpConstructor: op.constructDiskBackedOp,
 		spillingCallbackFn:      spillingCallbackFn,
 	}
@@ -219,7 +213,7 @@ type diskSpillerBase struct {
 	spilled bool
 
 	inMemoryOp              colexecop.BufferingInMemoryOperator
-	inMemoryMemMonitorNames []string
+	inMemoryMemMonitorNames []mon.Name
 	// diskBackedOp is created lazily when the diskSpillerBase spills to disk
 	// for the first time throughout its lifetime.
 	diskBackedOp            colexecop.Operator
@@ -256,7 +250,7 @@ func (d *diskSpillerBase) Next() coldata.Batch {
 			// Check if this error is from one of our memory monitors.
 			var found bool
 			for i := range d.inMemoryMemMonitorNames {
-				if strings.Contains(err.Error(), d.inMemoryMemMonitorNames[i]) {
+				if strings.Contains(err.Error(), d.inMemoryMemMonitorNames[i].String()) {
 					found = true
 					break
 				}