@mgartner 2025-05-13
Here's one benchmark showing the improvement:

```
name                                                                                                                old time/op    new time/op    delta
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int         22.5µs ± 0%    21.6µs ± 0%   -4.31%  (p=0.000 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes       44.8µs ± 1%    44.1µs ± 1%   -1.66%  (p=0.000 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int        1.24ms ± 0%    1.23ms ± 0%   -0.39%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes      2.36ms ± 1%    2.36ms ± 1%     ~     (p=0.143 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int         26.2µs ± 0%    25.3µs ± 1%   -3.32%  (p=0.000 n=9+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes       50.0µs ± 1%    49.5µs ± 0%   -0.98%  (p=0.009 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int        2.11ms ± 1%    2.11ms ± 0%     ~     (p=0.661 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes      3.54ms ± 1%    3.55ms ± 1%     ~     (p=0.165 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int       132ms ± 2%     131ms ± 0%     ~     (p=0.360 n=10+8)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes     237ms ± 3%     242ms ± 4%   +2.17%  (p=0.029 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int          3.99ms ± 1%    3.99ms ± 1%     ~     (p=1.000 n=10+9)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes        4.01ms ± 1%    4.03ms ± 1%     ~     (p=0.447 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int         13.3ms ± 1%    13.3ms ± 1%     ~     (p=0.218 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes       14.4ms ± 1%    14.5ms ± 2%     ~     (p=0.436 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int          11.8ms ± 1%    11.8ms ± 1%     ~     (p=0.497 n=10+9)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes        11.8ms ± 0%    11.9ms ± 1%     ~     (p=0.243 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int         23.9ms ± 9%    22.7ms ± 1%   -4.85%  (p=0.023 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes       28.2ms ± 2%    28.2ms ± 1%     ~     (p=0.661 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int        867ms ± 1%     870ms ± 1%     ~     (p=0.387 n=9+9)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes      1.86s ± 1%     1.90s ± 3%   +2.44%  (p=0.003 n=8+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int           8.02ms ± 1%    7.99ms ± 1%     ~     (p=0.393 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes         10.0ms ± 4%     8.0ms ± 1%  -20.02%  (p=0.000 n=8+8)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int          17.3ms ± 2%    17.4ms ± 1%     ~     (p=0.095 n=9+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes        18.4ms ± 0%    18.5ms ± 1%   +0.67%  (p=0.040 n=9+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int           15.8ms ± 1%    15.8ms ± 1%     ~     (p=0.356 n=10+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes         15.8ms ± 1%    15.9ms ± 1%   +0.74%  (p=0.006 n=10+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int          28.0ms ± 1%    28.1ms ± 1%     ~     (p=0.105 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes        34.2ms ± 2%    34.2ms ± 1%     ~     (p=0.780 n=10+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int         947ms ± 1%     950ms ± 2%     ~     (p=0.897 n=8+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes       2.04s ± 2%     2.03s ± 1%     ~     (p=0.089 n=10+10)

name                                                                                                                old speed      new speed      delta
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int        727MB/s ± 0%   760MB/s ± 0%   +4.51%  (p=0.000 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes      366MB/s ± 1%   372MB/s ± 1%   +1.69%  (p=0.000 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int       846MB/s ± 0%   849MB/s ± 0%   +0.40%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes     444MB/s ± 1%   445MB/s ± 1%     ~     (p=0.143 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int        626MB/s ± 0%   648MB/s ± 1%   +3.44%  (p=0.000 n=9+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes      328MB/s ± 1%   331MB/s ± 0%   +0.98%  (p=0.009 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int       496MB/s ± 1%   496MB/s ± 0%     ~     (p=0.661 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes     296MB/s ± 1%   295MB/s ± 1%     ~     (p=0.165 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int     510MB/s ± 2%   511MB/s ± 0%     ~     (p=0.360 n=10+8)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes   283MB/s ± 3%   277MB/s ± 3%   -2.09%  (p=0.029 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int        4.10MB/s ± 1%  4.11MB/s ± 0%     ~     (p=0.885 n=10+7)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes      4.08MB/s ± 1%  4.07MB/s ± 1%     ~     (p=0.531 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int       79.1MB/s ± 1%  78.8MB/s ± 1%     ~     (p=0.218 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes     72.7MB/s ± 1%  72.5MB/s ± 2%     ~     (p=0.446 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int        1.39MB/s ± 1%  1.39MB/s ± 0%     ~     (p=1.000 n=10+7)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes      1.38MB/s ± 1%  1.38MB/s ± 1%     ~     (p=0.473 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int       44.0MB/s ± 8%  46.1MB/s ± 1%   +4.80%  (p=0.022 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes     37.2MB/s ± 2%  37.2MB/s ± 1%     ~     (p=0.675 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int     77.4MB/s ± 1%  77.2MB/s ± 1%     ~     (p=0.352 n=9+9)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes   36.1MB/s ± 1%  35.2MB/s ± 3%   -2.36%  (p=0.003 n=8+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int         2.05MB/s ± 1%  2.05MB/s ± 1%     ~     (p=0.669 n=8+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes       1.64MB/s ± 4%  2.05MB/s ± 1%  +24.87%  (p=0.000 n=8+8)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int        60.6MB/s ± 2%  60.3MB/s ± 1%     ~     (p=0.095 n=9+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes      57.1MB/s ± 0%  56.7MB/s ± 1%   -0.66%  (p=0.034 n=9+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int         1.04MB/s ± 1%  1.03MB/s ± 1%     ~     (p=0.174 n=10+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes       1.04MB/s ± 1%  1.03MB/s ± 2%   -0.87%  (p=0.019 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int        37.5MB/s ± 1%  37.3MB/s ± 1%     ~     (p=0.100 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes      30.7MB/s ± 2%  30.7MB/s ± 1%     ~     (p=0.734 n=10+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int      70.9MB/s ± 1%  70.6MB/s ± 2%     ~     (p=0.897 n=8+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes    32.8MB/s ± 2%  33.0MB/s ± 1%     ~     (p=0.085 n=10+10)

name                                                                                                                old alloc/op   new alloc/op   delta
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int         54.6kB ± 0%    54.5kB ± 0%   -0.25%  (p=0.000 n=9+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes        104kB ± 0%     104kB ± 0%   -0.13%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int         101kB ± 0%     101kB ± 0%   -0.14%  (p=0.000 n=9+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes       168kB ± 0%     167kB ± 0%   -0.08%  (p=0.000 n=9+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int         59.2kB ± 0%    59.0kB ± 0%   -0.23%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes        110kB ± 0%     110kB ± 0%   -0.12%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int         629kB ± 0%     628kB ± 0%   -0.02%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes      1.17MB ± 0%    1.17MB ± 0%   -0.01%  (p=0.002 n=8+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int      49.6MB ± 0%    49.6MB ± 0%   -0.00%  (p=0.000 n=10+9)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes    83.1MB ± 0%    83.1MB ± 0%   -0.01%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int           162kB ± 0%     162kB ± 0%   -0.12%  (p=0.000 n=8+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes         360kB ± 0%     359kB ± 0%   -0.06%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int          497kB ± 0%     497kB ± 0%   -0.04%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes        644kB ± 0%     644kB ± 0%   -0.04%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int           209kB ± 0%     209kB ± 0%   -0.10%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes         437kB ± 0%     437kB ± 0%   -0.04%  (p=0.001 n=10+8)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int         1.34MB ± 0%    1.34MB ± 0%     ~     (p=0.436 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes       2.83MB ± 0%    2.83MB ± 0%   -0.02%  (p=0.001 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int       32.4MB ± 0%    32.4MB ± 0%   -0.00%  (p=0.006 n=10+8)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes     72.3MB ± 1%    71.8MB ± 2%     ~     (p=0.116 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int            216kB ± 0%     216kB ± 0%   -0.09%  (p=0.000 n=8+8)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes          464kB ± 0%     464kB ± 0%   -0.04%  (p=0.000 n=8+8)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int           562kB ± 0%     562kB ± 0%   -0.04%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes         794kB ± 0%     793kB ± 0%   -0.03%  (p=0.010 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int            283kB ± 0%     282kB ± 0%   -0.07%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes          594kB ± 0%     593kB ± 0%     ~     (p=0.579 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int          2.11MB ± 0%    2.11MB ± 0%     ~     (p=0.633 n=8+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes        4.93MB ± 0%    4.93MB ± 0%     ~     (p=0.393 n=10+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int        79.8MB ± 0%    79.8MB ± 0%     ~     (p=0.952 n=9+10)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes       166MB ±16%     163MB ±14%     ~     (p=0.182 n=10+10)

name                                                                                                                old allocs/op  new allocs/op  delta
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int           71.0 ± 0%      65.0 ± 0%   -8.45%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes         67.0 ± 0%      61.0 ± 0%   -8.96%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int           594 ± 0%       588 ± 0%   -1.01%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes         574 ± 0%       568 ± 0%   -1.05%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int           85.0 ± 0%      79.0 ± 0%   -7.06%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes         67.0 ± 0%      61.0 ± 0%   -8.96%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int           752 ± 0%       746 ± 0%   -0.80%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes         598 ± 0%       592 ± 0%   -1.00%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int       41.2k ± 0%     41.2k ± 0%   -0.02%  (p=0.000 n=10+10)
ExternalDistinct/spilled=false/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes     32.9k ± 0%     32.9k ± 0%   -0.03%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int             415 ± 0%       408 ± 0%   -1.69%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=bytes           410 ± 0%       403 ± 0%   -1.71%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=int          2.28k ± 0%     2.27k ± 0%   -0.29%  (p=0.000 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.001/rows=65536/ordCols=0/type=bytes        1.79k ± 0%     1.78k ± 0%   -0.40%  (p=0.000 n=10+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=int             667 ± 0%       660 ± 0%   -1.05%  (p=0.000 n=8+8)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=1024/ordCols=0/type=bytes           645 ± 0%       638 ± 0%   -1.05%  (p=0.000 n=9+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=int          3.54k ± 0%     3.53k ± 0%   -0.20%  (p=0.000 n=6+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=65536/ordCols=0/type=bytes        4.04k ± 0%     4.04k ± 0%   -0.19%  (p=0.000 n=8+10)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=int         193k ± 0%      193k ± 0%   -0.00%  (p=0.000 n=10+8)
ExternalDistinct/spilled=true/ordering=false/shuffled/newTupleProbability=0.100/rows=4194304/ordCols=0/type=bytes       553k ± 0%      553k ± 0%   -0.00%  (p=0.007 n=7+9)
ExternalDistinct/spilled=true/ordering=true/shuffled/newTupleProbability=0.001/rows=1024/ordCols=0/type=int              748 ± 0%       740 ± 0%   -1.07%  (p=0.000 n=10+10)
…